Moves NavBar and Footer up one level and updates imports accordingly
