+++ src/components/Navbar.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { NavLink, useLocation } from 'react-router-dom'
-import { IconSmallPlus } from '../common/IconSmallPlus'
+import { IconSmallPlus } from './common/IconSmallPlus'
 
 const navLinksParams = {
   about: {
