+++ src/index.js
@@ -1,8 +1,8 @@
 import { BrowserRouter, Route, Redirect, Switch } from "react-router-dom"
 import React from "react"
 import ReactDOM from "react-dom"
-import { Footer } from "./components/layouts/Footer"
-import { NavBar } from "./components/layouts/Navbar"
+import { Footer } from "./components/Footer"
+import { NavBar } from "./components/Navbar"
 import { About } from "./components/about/About"
 import { Portfolio } from "./components/portfolio/Portfolio"
 import { Resume } from "./components/resume/Resume"
